transfers: reduce stutter in naming
