+++ internal/transfers/transfers.go
@@ -12,7 +12,6 @@ import (
 	"io/ioutil"
 	"net/http"
 	"strings"
-	"time"
 
 	moovaccounts "github.com/moov-io/accounts/client"
 	"github.com/moov-io/ach"
@@ -20,7 +19,6 @@ import (
 	"github.com/moov-io/base"
 	"github.com/moov-io/base/idempotent"
 	moovcustomers "github.com/moov-io/customers"
-	"github.com/moov-io/paygate/internal"
 	"github.com/moov-io/paygate/internal/accounts"
 	"github.com/moov-io/paygate/internal/customers"
 	"github.com/moov-io/paygate/internal/depository"
@@ -119,7 +117,7 @@ type TransferRouter struct {
 	eventRepo          events.Repository
 	receiverRepository receivers.Repository
 	origRepo           originators.Repository
-	transferRepo       TransferRepository
+	transferRepo       Repository
 
 	transferLimitChecker *LimitChecker
 
@@ -135,7 +133,7 @@ func NewTransferRouter(
 	eventRepo events.Repository,
 	receiverRepo receivers.Repository,
 	originatorsRepo originators.Repository,
-	transferRepo TransferRepository,
+	transferRepo Repository,
 	transferLimitChecker *LimitChecker,
 	achClientFactory func(userID id.User) *achclient.ACH,
 	accountsClient accounts.Client,
@@ -556,353 +554,6 @@ func (c *TransferRouter) getUserTransferEvents() http.HandlerFunc {
 			json.NewEncoder(w).Encode(events)
 		})
 	}
-}
-
-type TransferRepository interface {
-	getUserTransfers(userID id.User) ([]*model.Transfer, error)
-	getUserTransfer(id id.Transfer, userID id.User) (*model.Transfer, error)
-	UpdateTransferStatus(id id.Transfer, status model.TransferStatus) error
-
-	GetFileIDForTransfer(id id.Transfer, userID id.User) (string, error)
-
-	LookupTransferFromReturn(sec string, amount *model.Amount, traceNumber string, effectiveEntryDate time.Time) (*model.Transfer, error)
-	SetReturnCode(id id.Transfer, returnCode string) error
-
-	// GetTransferCursor returns a database cursor for Transfer objects that need to be
-	// posted today.
-	//
-	// We currently default EffectiveEntryDate to tomorrow for any transfer and thus a
-	// transfer created today needs to be posted.
-	GetTransferCursor(batchSize int, depRepo depository.Repository) *TransferCursor
-	MarkTransferAsMerged(id id.Transfer, filename string, traceNumber string) error
-
-	createUserTransfers(userID id.User, requests []*transferRequest) ([]*model.Transfer, error)
-	deleteUserTransfer(id id.Transfer, userID id.User) error
-}
-
-func NewTransferRepo(logger log.Logger, db *sql.DB) *SQLTransferRepo {
-	return &SQLTransferRepo{log: logger, db: db}
-}
-
-type SQLTransferRepo struct {
-	db  *sql.DB
-	log log.Logger
-}
-
-func (r *SQLTransferRepo) Close() error {
-	return r.db.Close()
-}
-
-func (r *SQLTransferRepo) getUserTransfers(userID id.User) ([]*model.Transfer, error) {
-	query := `select transfer_id from transfers where user_id = ? and deleted_at is null`
-	stmt, err := r.db.Prepare(query)
-	if err != nil {
-		return nil, err
-	}
-	defer stmt.Close()
-
-	rows, err := stmt.Query(userID)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-
-	var transferIDs []string
-	for rows.Next() {
-		var row string
-		if err := rows.Scan(&row); err != nil {
-			return nil, fmt.Errorf("getUserTransfers scan: %v", err)
-		}
-		if row != "" {
-			transferIDs = append(transferIDs, row)
-		}
-	}
-	if err := rows.Err(); err != nil {
-		return nil, fmt.Errorf("getUserTransfers: rows.Err=%v", err)
-	}
-
-	var transfers []*model.Transfer
-	for i := range transferIDs {
-		t, err := r.getUserTransfer(id.Transfer(transferIDs[i]), userID)
-		if err == nil && t.ID != "" {
-			transfers = append(transfers, t)
-		}
-	}
-	return transfers, rows.Err()
-}
-
-func (r *SQLTransferRepo) getUserTransfer(id id.Transfer, userID id.User) (*model.Transfer, error) {
-	query := `select transfer_id, type, amount, originator_id, originator_depository, receiver, receiver_depository, description, standard_entry_class_code, status, same_day, return_code, created_at
-from transfers
-where transfer_id = ? and user_id = ? and deleted_at is null
-limit 1`
-	stmt, err := r.db.Prepare(query)
-	if err != nil {
-		return nil, err
-	}
-	defer stmt.Close()
-
-	row := stmt.QueryRow(id, userID)
-
-	transfer := &model.Transfer{}
-	var (
-		amt        string
-		returnCode *string
-		created    time.Time
-	)
-	err = row.Scan(&transfer.ID, &transfer.Type, &amt, &transfer.Originator, &transfer.OriginatorDepository, &transfer.Receiver, &transfer.ReceiverDepository, &transfer.Description, &transfer.StandardEntryClassCode, &transfer.Status, &transfer.SameDay, &returnCode, &created)
-	if err != nil {
-		return nil, err
-	}
-	if returnCode != nil {
-		transfer.ReturnCode = ach.LookupReturnCode(*returnCode)
-	}
-	transfer.Created = base.NewTime(created)
-	// parse Amount struct
-	if err := transfer.Amount.FromString(amt); err != nil {
-		return nil, err
-	}
-	if transfer.ID == "" {
-		return nil, nil // not found
-	}
-	return transfer, nil
-}
-
-func (r *SQLTransferRepo) UpdateTransferStatus(id id.Transfer, status model.TransferStatus) error {
-	query := `update transfers set status = ? where transfer_id = ? and deleted_at is null`
-	stmt, err := r.db.Prepare(query)
-	if err != nil {
-		return err
-	}
-	defer stmt.Close()
-
-	_, err = stmt.Exec(status, id)
-	return err
-}
-
-func (r *SQLTransferRepo) GetFileIDForTransfer(id id.Transfer, userID id.User) (string, error) {
-	query := `select file_id from transfers where transfer_id = ? and user_id = ? and deleted_at is null limit 1;`
-	stmt, err := r.db.Prepare(query)
-	if err != nil {
-		return "", err
-	}
-	defer stmt.Close()
-
-	row := stmt.QueryRow(id, userID)
-
-	var fileID string
-	if err := row.Scan(&fileID); err != nil {
-		return "", err
-	}
-	return fileID, nil
-}
-
-func (r *SQLTransferRepo) LookupTransferFromReturn(sec string, amount *model.Amount, traceNumber string, effectiveEntryDate time.Time) (*model.Transfer, error) {
-	// To match returned files we take a few values which are assumed to uniquely identify a Transfer.
-	// traceNumber, per NACHA guidelines, should be globally unique (routing number + random value),
-	// but we are going to filter to only select Transfers created within a few days of the EffectiveEntryDate
-	// to avoid updating really old (or future, I suppose) objects.
-	query := `select transfer_id, user_id, transaction_id from transfers
-where standard_entry_class_code = ? and amount = ? and trace_number = ? and status = ? and (created_at > ? and created_at < ?) and deleted_at is null limit 1`
-	stmt, err := r.db.Prepare(query)
-	if err != nil {
-		return nil, err
-	}
-	defer stmt.Close()
-
-	transferId, userID, transactionID := "", "", "" // holders for 'select ..'
-	min, max := internal.StartOfDayAndTomorrow(effectiveEntryDate)
-	// Only include Transfer objects within 5 calendar days of the EffectiveEntryDate
-	min = min.Add(-5 * 24 * time.Hour)
-	max = max.Add(5 * 24 * time.Hour)
-
-	row := stmt.QueryRow(sec, amount.String(), traceNumber, model.TransferProcessed, min, max)
-	if err := row.Scan(&transferId, &userID, &transactionID); err != nil {
-		return nil, err
-	}
-
-	xfer, err := r.getUserTransfer(id.Transfer(transferId), id.User(userID))
-	xfer.TransactionID = transactionID
-	xfer.UserID = userID
-	return xfer, err
-}
-
-func (r *SQLTransferRepo) SetReturnCode(id id.Transfer, returnCode string) error {
-	query := `update transfers set return_code = ? where transfer_id = ? and return_code is null and deleted_at is null`
-	stmt, err := r.db.Prepare(query)
-	if err != nil {
-		return err
-	}
-	defer stmt.Close()
-
-	_, err = stmt.Exec(returnCode, id)
-	return err
-}
-
-func (r *SQLTransferRepo) createUserTransfers(userID id.User, requests []*transferRequest) ([]*model.Transfer, error) {
-	query := `insert into transfers (transfer_id, user_id, type, amount, originator_id, originator_depository, receiver, receiver_depository, description, standard_entry_class_code, status, same_day, file_id, transaction_id, remote_address, created_at) values (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`
-	stmt, err := r.db.Prepare(query)
-	if err != nil {
-		return nil, err
-	}
-	defer stmt.Close()
-
-	var transfers []*model.Transfer
-
-	now := time.Now()
-	var status model.TransferStatus = model.TransferPending
-	for i := range requests {
-		req, transferId := requests[i], base.ID()
-		xfer := &model.Transfer{
-			ID:                     id.Transfer(transferId),
-			Type:                   req.Type,
-			Amount:                 req.Amount,
-			Originator:             req.Originator,
-			OriginatorDepository:   req.OriginatorDepository,
-			Receiver:               req.Receiver,
-			ReceiverDepository:     req.ReceiverDepository,
-			Description:            req.Description,
-			StandardEntryClassCode: req.StandardEntryClassCode,
-			Status:                 status,
-			SameDay:                req.SameDay,
-			Created:                base.NewTime(now),
-		}
-		if err := xfer.Validate(); err != nil {
-			return nil, fmt.Errorf("validation failed for transfer Originator=%s, Receiver=%s, Description=%s %v", xfer.Originator, xfer.Receiver, xfer.Description, err)
-		}
-
-		// write transfer
-		_, err := stmt.Exec(transferId, userID, req.Type, req.Amount.String(), req.Originator, req.OriginatorDepository, req.Receiver, req.ReceiverDepository, req.Description, req.StandardEntryClassCode, status, req.SameDay, req.fileID, req.transactionID, req.remoteAddr, now)
-		if err != nil {
-			return nil, err
-		}
-		transfers = append(transfers, xfer)
-	}
-	return transfers, nil
-}
-
-func (r *SQLTransferRepo) deleteUserTransfer(id id.Transfer, userID id.User) error {
-	query := `update transfers set deleted_at = ? where transfer_id = ? and user_id = ? and deleted_at is null`
-	stmt, err := r.db.Prepare(query)
-	if err != nil {
-		return err
-	}
-	defer stmt.Close()
-
-	_, err = stmt.Exec(time.Now(), id, userID)
-	return err
-}
-
-// TransferCursor allows for iterating through Transfers in ascending order (by CreatedAt)
-// to merge into files uploaded to an ODFI.
-type TransferCursor struct {
-	BatchSize int
-
-	DepRepo      depository.Repository
-	TransferRepo *SQLTransferRepo
-
-	// newerThan represents the minimum (oldest) created_at value to return in the batch.
-	// The value starts at today's first instant and progresses towards time.Now() with each
-	// batch by being set to the batch's newest time.
-	newerThan time.Time
-}
-
-// GroupableTransfer holds metadata of a Transfer used in grouping for generating and merging ACH files
-// to be uploaded into the Fed.
-type GroupableTransfer struct {
-	*model.Transfer
-
-	// Destination is the ABA routing number of the Destination FI (DFI)
-	// This comes from the Transfer's ReceiverDepository.RoutingNumber
-	Destination string
-
-	UserID id.User
-}
-
-// Next returns a slice of Transfer objects from the current day. Next should be called to process
-// all objects for a given day in batches.
-//
-// TODO(adam): should we have a field on transfers for marking when the ACH file is uploaded?
-// "after the file is uploaded we mark the items in the DB with the batch number and upload time and update the status" -- Wade
-func (cur *TransferCursor) Next() ([]*GroupableTransfer, error) {
-	query := `select transfer_id, user_id, created_at from transfers where status = ? and merged_filename is null and created_at > ? and deleted_at is null order by created_at asc limit ?`
-	stmt, err := cur.TransferRepo.db.Prepare(query)
-	if err != nil {
-		return nil, fmt.Errorf("TransferCursor.Next: prepare: %v", err)
-	}
-	defer stmt.Close()
-
-	rows, err := stmt.Query(model.TransferPending, cur.newerThan, cur.BatchSize) // only Pending transfers
-	if err != nil {
-		return nil, fmt.Errorf("TransferCursor.Next: query: %v", err)
-	}
-	defer rows.Close()
-
-	type xfer struct {
-		transferId, userID string
-		createdAt          time.Time
-	}
-	var xfers []xfer
-	for rows.Next() {
-		var xf xfer
-		if err := rows.Scan(&xf.transferId, &xf.userID, &xf.createdAt); err != nil {
-			return nil, fmt.Errorf("TransferCursor.Next: scan: %v", err)
-		}
-		if xf.transferId != "" {
-			xfers = append(xfers, xf)
-		}
-	}
-
-	max := cur.newerThan
-
-	var transfers []*GroupableTransfer
-	for i := range xfers {
-		t, err := cur.TransferRepo.getUserTransfer(id.Transfer(xfers[i].transferId), id.User(xfers[i].userID))
-		if err != nil {
-			continue
-		}
-		destDep, err := cur.DepRepo.GetUserDepository(t.ReceiverDepository, id.User(xfers[i].userID))
-		if err != nil || destDep == nil {
-			continue
-		}
-		transfers = append(transfers, &GroupableTransfer{
-			Transfer:    t,
-			Destination: destDep.RoutingNumber,
-			UserID:      id.User(xfers[i].userID),
-		})
-		if xfers[i].createdAt.After(max) {
-			max = xfers[i].createdAt // advance max to newest time
-		}
-	}
-	cur.newerThan = max
-	return transfers, rows.Err()
-}
-
-// GetTransferCursor returns a TransferCursor for iterating through Transfers in ascending order (by CreatedAt)
-// beginning at the start of the current day.
-func (r *SQLTransferRepo) GetTransferCursor(batchSize int, depRepo depository.Repository) *TransferCursor {
-	now := time.Now()
-	return &TransferCursor{
-		BatchSize:    batchSize,
-		TransferRepo: r,
-		newerThan:    time.Date(now.Year(), now.Month(), now.Day(), 0, 0, 0, 0, time.UTC),
-		DepRepo:      depRepo,
-	}
-}
-
-// MarkTransferAsMerged will set the merged_filename on Pending transfers so they aren't merged into multiple files
-// and the file uploaded to the FED can be tracked.
-func (r *SQLTransferRepo) MarkTransferAsMerged(id id.Transfer, filename string, traceNumber string) error {
-	query := `update transfers set merged_filename = ?, trace_number = ?, status = ?
-where status = ? and transfer_id = ? and (merged_filename is null or merged_filename = '') and deleted_at is null`
-	stmt, err := r.db.Prepare(query)
-	if err != nil {
-		return fmt.Errorf("MarkTransferAsMerged: transfer=%s filename=%s: %v", id, filename, err)
-	}
-	defer stmt.Close()
-
-	_, err = stmt.Exec(filename, traceNumber, model.TransferProcessed, model.TransferPending, id)
-	return err
 }
 
 // getTransferObjects performs database lookups to grab all the objects needed to make a transfer.
