+++ internal/transfers/transfers_test.go
@@ -57,14 +57,14 @@ func CreateTestTransferRouter(
 	evt events.Repository,
 	rec receivers.Repository,
 	ori originators.Repository,
-	xfr TransferRepository,
+	xfr Repository,
 	routes ...func(*mux.Router), // test ACH server routes
 ) *testTransferRouter {
 
 	limits, _ := ParseLimits(SevenDayLimit(), ThirtyDayLimit())
 
 	var db *sql.DB
-	if rr, ok := xfr.(*SQLTransferRepo); ok {
+	if rr, ok := xfr.(*SQLRepo); ok {
 		db = rr.db
 	}
 	limiter := NewLimitChecker(log.NewNopLogger(), db, limits)
@@ -497,7 +497,7 @@ func TestTransfers__create(t *testing.T) {
 			},
 		},
 	}
-	repo := &SQLTransferRepo{db.DB, log.NewNopLogger()}
+	repo := &SQLRepo{db.DB, log.NewNopLogger()}
 
 	amt, _ := model.NewAmount("USD", "18.61")
 	request := &transferRequest{
@@ -578,7 +578,7 @@ func TestTransfers__getUserTransfer(t *testing.T) {
 	db := database.CreateTestSqliteDB(t)
 	defer db.Close()
 
-	repo := &SQLTransferRepo{db.DB, log.NewNopLogger()}
+	repo := &SQLRepo{db.DB, log.NewNopLogger()}
 
 	amt, _ := model.NewAmount("USD", "18.61")
 	userID := id.User(base.ID())
@@ -636,7 +636,7 @@ func TestTransfers__getUserTransfer(t *testing.T) {
 	}
 
 	// have our repository error and verify we get non-200's
-	xferRouter.transferRepo = &MockTransferRepository{Err: errors.New("bad error")}
+	xferRouter.transferRepo = &MockRepository{Err: errors.New("bad error")}
 
 	w = httptest.NewRecorder()
 	router.ServeHTTP(w, r)
@@ -651,7 +651,7 @@ func TestTransfers__getUserTransfers(t *testing.T) {
 	db := database.CreateTestSqliteDB(t)
 	defer db.Close()
 
-	repo := &SQLTransferRepo{db.DB, log.NewNopLogger()}
+	repo := &SQLRepo{db.DB, log.NewNopLogger()}
 
 	amt, _ := model.NewAmount("USD", "12.42")
 	userID := id.User(base.ID())
@@ -707,7 +707,7 @@ func TestTransfers__getUserTransfers(t *testing.T) {
 	}
 
 	// have our repository error and verify we get non-200's
-	xferRouter.transferRepo = &MockTransferRepository{Err: errors.New("bad error")}
+	xferRouter.transferRepo = &MockRepository{Err: errors.New("bad error")}
 
 	w = httptest.NewRecorder()
 	router.ServeHTTP(w, r)
@@ -722,7 +722,7 @@ func TestTransfers__deleteUserTransfer(t *testing.T) {
 	db := database.CreateTestSqliteDB(t)
 	defer db.Close()
 
-	repo := &SQLTransferRepo{db.DB, log.NewNopLogger()}
+	repo := &SQLRepo{db.DB, log.NewNopLogger()}
 
 	amt, _ := model.NewAmount("USD", "12.42")
 	userID := id.User(base.ID())
@@ -763,7 +763,7 @@ func TestTransfers__deleteUserTransfer(t *testing.T) {
 	}
 
 	// have our repository error and verify we get non-200's
-	xferRouter.transferRepo = &MockTransferRepository{Err: errors.New("bad error")}
+	xferRouter.transferRepo = &MockRepository{Err: errors.New("bad error")}
 
 	w = httptest.NewRecorder()
 	router.ServeHTTP(w, r)
@@ -778,7 +778,7 @@ func TestTransfers__validateUserTransfer(t *testing.T) {
 	db := database.CreateTestSqliteDB(t)
 	defer db.Close()
 
-	repo := &SQLTransferRepo{db.DB, log.NewNopLogger()}
+	repo := &SQLRepo{db.DB, log.NewNopLogger()}
 
 	amt, _ := model.NewAmount("USD", "32.41")
 	userID := id.User(base.ID())
@@ -818,7 +818,7 @@ func TestTransfers__validateUserTransfer(t *testing.T) {
 	}
 
 	// have our repository error and verify we get non-200's
-	mockRepo := &MockTransferRepository{Err: errors.New("bad error")}
+	mockRepo := &MockRepository{Err: errors.New("bad error")}
 	xferRouter.transferRepo = mockRepo
 
 	w = httptest.NewRecorder()
@@ -848,7 +848,7 @@ func TestTransfers__getUserTransferFiles(t *testing.T) {
 	db := database.CreateTestSqliteDB(t)
 	defer db.Close()
 
-	repo := &SQLTransferRepo{db.DB, log.NewNopLogger()}
+	repo := &SQLRepo{db.DB, log.NewNopLogger()}
 
 	amt, _ := model.NewAmount("USD", "32.41")
 	userID := id.User(base.ID())
@@ -902,7 +902,7 @@ func TestTransfers__getUserTransferFiles(t *testing.T) {
 	}
 
 	// have our repository error and verify we get non-200's
-	mockRepo := &MockTransferRepository{Err: errors.New("bad error")}
+	mockRepo := &MockRepository{Err: errors.New("bad error")}
 	xferRouter.transferRepo = mockRepo
 
 	w = httptest.NewRecorder()
@@ -961,217 +961,6 @@ func TestTransfers__writeResponse(t *testing.T) {
 	}
 }
 
-func TestTransfers_transferCursor(t *testing.T) {
-	db := database.CreateTestSqliteDB(t)
-	defer db.Close()
-
-	keeper := secrets.TestStringKeeper(t)
-
-	depRepo := depository.NewDepositoryRepo(log.NewNopLogger(), db.DB, keeper)
-	transferRepo := &SQLTransferRepo{db.DB, log.NewNopLogger()}
-
-	userID := id.User(base.ID())
-	amt := func(number string) model.Amount {
-		amt, _ := model.NewAmount("USD", number)
-		return *amt
-	}
-
-	dep := &model.Depository{
-		ID:                     id.Depository(base.ID()),
-		BankName:               "bank name",
-		Holder:                 "holder",
-		HolderType:             model.Individual,
-		Type:                   model.Checking,
-		RoutingNumber:          "123",
-		EncryptedAccountNumber: "151",
-		Status:                 model.DepositoryUnverified,
-		Created:                base.NewTime(time.Now().Add(-1 * time.Second)),
-	}
-	if err := depRepo.UpsertUserDepository(userID, dep); err != nil {
-		t.Fatal(err)
-	}
-
-	// Write transfers into database
-	//
-	// With the batch size low enough if more transfers are inserted within 1ms? than the batch size
-	// the cursor will get stuck in an infinite loop. So we're inserting them at different times.
-	//
-	// TODO(adam): Will this become an issue?
-	requests := []*transferRequest{
-		{
-			Type:                   model.PushTransfer,
-			Amount:                 amt("12.12"),
-			Originator:             model.OriginatorID("originator1"),
-			OriginatorDepository:   dep.ID, // OriginatorDepository is read from a depositoryRepository
-			Receiver:               model.ReceiverID("receiver1"),
-			ReceiverDepository:     id.Depository("receiver1"),
-			Description:            "money1",
-			StandardEntryClassCode: "PPD",
-			fileID:                 "test-file1",
-		},
-	}
-	if _, err := transferRepo.createUserTransfers(userID, requests); err != nil {
-		t.Fatal(err)
-	}
-	requests = []*transferRequest{
-		{
-			Type:                   model.PullTransfer,
-			Amount:                 amt("13.13"),
-			Originator:             model.OriginatorID("originator2"),
-			OriginatorDepository:   dep.ID,
-			Receiver:               model.ReceiverID("receiver2"),
-			ReceiverDepository:     id.Depository("receiver2"),
-			Description:            "money2",
-			StandardEntryClassCode: "PPD",
-			fileID:                 "test-file2",
-		},
-	}
-	if _, err := transferRepo.createUserTransfers(userID, requests); err != nil {
-		t.Fatal(err)
-	}
-	requests = []*transferRequest{
-		{
-			Type:                   model.PushTransfer,
-			Amount:                 amt("14.14"),
-			Originator:             model.OriginatorID("originator3"),
-			OriginatorDepository:   dep.ID,
-			Receiver:               model.ReceiverID("receiver3"),
-			ReceiverDepository:     id.Depository("receiver3"),
-			Description:            "money3",
-			StandardEntryClassCode: "PPD",
-			fileID:                 "test-file3",
-		},
-	}
-	if _, err := transferRepo.createUserTransfers(userID, requests); err != nil {
-		t.Fatal(err)
-	}
-
-	// Now verify the cursor pulls those transfers out
-	cur := transferRepo.GetTransferCursor(2, depRepo) // batch size
-	firstBatch, err := cur.Next()
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(firstBatch) != 2 {
-		for i := range firstBatch {
-			t.Errorf("firstBatch[%d]=%#v", i, firstBatch[i])
-		}
-	}
-	secondBatch, err := cur.Next()
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(secondBatch) != 1 {
-		for i := range secondBatch {
-			t.Errorf("secondBatch[%d]=%#v", i, secondBatch[i])
-		}
-	}
-}
-
-func TestTransfers_MarkTransferAsMerged(t *testing.T) {
-	db := database.CreateTestSqliteDB(t)
-	defer db.Close()
-
-	keeper := secrets.TestStringKeeper(t)
-	depRepo := depository.NewDepositoryRepo(log.NewNopLogger(), db.DB, keeper)
-	transferRepo := &SQLTransferRepo{db.DB, log.NewNopLogger()}
-
-	userID := id.User(base.ID())
-	amt := func(number string) model.Amount {
-		amt, _ := model.NewAmount("USD", number)
-		return *amt
-	}
-
-	dep := &model.Depository{
-		ID:                     id.Depository(base.ID()),
-		BankName:               "bank name",
-		Holder:                 "holder",
-		HolderType:             model.Individual,
-		Type:                   model.Checking,
-		RoutingNumber:          "123",
-		EncryptedAccountNumber: "151",
-		Status:                 model.DepositoryVerified,
-		Created:                base.NewTime(time.Now().Add(-1 * time.Second)),
-	}
-	if err := depRepo.UpsertUserDepository(userID, dep); err != nil {
-		t.Fatal(err)
-	}
-
-	// Write transfers into database
-	//
-	// With the batch size low enough if more transfers are inserted within 1ms? than the batch size
-	// the cursor will get stuck in an infinite loop. So we're inserting them at different times.
-	//
-	// TODO(adam): Will this become an issue?
-	requests := []*transferRequest{
-		{
-			Type:                   model.PushTransfer,
-			Amount:                 amt("12.12"),
-			Originator:             model.OriginatorID("originator1"),
-			OriginatorDepository:   id.Depository("originator1"),
-			Receiver:               model.ReceiverID("receiver1"),
-			ReceiverDepository:     dep.ID, // ReceiverDepository is read from a depositoryRepository
-			Description:            "money1",
-			StandardEntryClassCode: "PPD",
-			fileID:                 "test-file1",
-		},
-	}
-	if _, err := transferRepo.createUserTransfers(userID, requests); err != nil {
-		t.Fatal(err)
-	}
-
-	// Now verify the cursor pulls those transfers out
-	cur := transferRepo.GetTransferCursor(2, depRepo) // batch size
-	firstBatch, err := cur.Next()
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(firstBatch) != 1 {
-		for i := range firstBatch {
-			t.Errorf("firstBatch[%d]=%#v", i, firstBatch[i])
-		}
-		t.Fatalf("firstBatch: %#v", firstBatch)
-	}
-
-	// mark our transfer as merged, so we don't see it (in a new transferCursor we create)
-	if err := transferRepo.MarkTransferAsMerged(firstBatch[0].ID, "merged-file.ach", "traceNumber"); err != nil {
-		t.Fatal(err)
-	}
-
-	// re-create our transferCursor and see the transfer ignored
-	// plus add a second transfer and ensure we get that
-	requests = []*transferRequest{
-		{
-			Type:                   model.PullTransfer,
-			Amount:                 amt("13.13"),
-			Originator:             model.OriginatorID("originator2"),
-			OriginatorDepository:   id.Depository("originator2"),
-			Receiver:               model.ReceiverID("receiver2"),
-			ReceiverDepository:     dep.ID,
-			Description:            "money2",
-			StandardEntryClassCode: "PPD",
-			fileID:                 "test-file2",
-		},
-	}
-	if _, err := transferRepo.createUserTransfers(userID, requests); err != nil {
-		t.Fatal(err)
-	}
-	cur = transferRepo.GetTransferCursor(2, depRepo) // batch size
-	firstBatch, err = cur.Next()
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(firstBatch) != 1 {
-		for i := range firstBatch {
-			t.Errorf("firstBatch[%d].ID=%v amount=%v", i, firstBatch[i].ID, firstBatch[i].Amount.String())
-		}
-		t.Fatalf("firstBatch: %#v", firstBatch)
-	}
-	if firstBatch[0].Amount.String() != "USD 13.13" {
-		t.Errorf("got %v", firstBatch[0].Amount.String())
-	}
-}
-
 func TestTransfers__createTransactionLines(t *testing.T) {
 	orig := &moovaccounts.Account{ID: base.ID()}
 	rec := &moovaccounts.Account{ID: base.ID()}
@@ -1215,7 +1004,7 @@ func TestTransfers__createTransactionLines(t *testing.T) {
 }
 
 func TestTransfers__postAccountTransaction(t *testing.T) {
-	transferRepo := &MockTransferRepository{}
+	transferRepo := &MockRepository{}
 
 	xferRouter := CreateTestTransferRouter(nil, nil, nil, nil, transferRepo)
 	defer xferRouter.close()
@@ -1251,220 +1040,6 @@ func TestTransfers__postAccountTransaction(t *testing.T) {
 	if tx == nil {
 		t.Errorf("nil accounts.Transaction")
 	}
-}
-
-func TestTransfers__UpdateTransferStatus(t *testing.T) {
-	t.Parallel()
-
-	check := func(t *testing.T, repo TransferRepository) {
-		amt, _ := model.NewAmount("USD", "32.92")
-		userID := id.User(base.ID())
-		req := &transferRequest{
-			Type:                   model.PushTransfer,
-			Amount:                 *amt,
-			Originator:             model.OriginatorID("originator"),
-			OriginatorDepository:   id.Depository("originator"),
-			Receiver:               model.ReceiverID("receiver"),
-			ReceiverDepository:     id.Depository("receiver"),
-			Description:            "money",
-			StandardEntryClassCode: "PPD",
-			fileID:                 "test-file",
-		}
-		transfers, err := repo.createUserTransfers(userID, []*transferRequest{req})
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		if err := repo.UpdateTransferStatus(transfers[0].ID, model.TransferReclaimed); err != nil {
-			t.Fatal(err)
-		}
-
-		xfer, err := repo.getUserTransfer(transfers[0].ID, userID)
-		if err != nil {
-			t.Error(err)
-		}
-		if xfer.Status != model.TransferReclaimed {
-			t.Errorf("got status %s", xfer.Status)
-		}
-	}
-
-	// SQLite tests
-	sqliteDB := database.CreateTestSqliteDB(t)
-	defer sqliteDB.Close()
-	check(t, &SQLTransferRepo{sqliteDB.DB, log.NewNopLogger()})
-
-	// MySQL tests
-	mysqlDB := database.CreateTestMySQLDB(t)
-	defer mysqlDB.Close()
-	check(t, &SQLTransferRepo{mysqlDB.DB, log.NewNopLogger()})
-}
-
-func TestTransfers__transactionID(t *testing.T) {
-	t.Parallel()
-
-	check := func(t *testing.T, db *sql.DB) {
-		userID := id.User(base.ID())
-		transactionID := base.ID() // field we care about
-		amt, _ := model.NewAmount("USD", "51.21")
-
-		repo := &SQLTransferRepo{db, log.NewNopLogger()}
-		requests := []*transferRequest{
-			{
-				Type:                   model.PullTransfer,
-				Amount:                 *amt,
-				Originator:             model.OriginatorID("originator"),
-				OriginatorDepository:   id.Depository("originatorDep"),
-				Receiver:               model.ReceiverID("receiver"),
-				ReceiverDepository:     id.Depository("receiverDep"),
-				Description:            "money2",
-				StandardEntryClassCode: "PPD",
-				transactionID:          transactionID,
-			},
-		}
-		if _, err := repo.createUserTransfers(userID, requests); err != nil {
-			t.Fatal(err)
-		}
-
-		transfers, err := repo.getUserTransfers(userID)
-		if err != nil || len(transfers) != 1 {
-			t.Errorf("got %d Transfers (error=%v): %v", len(transfers), err, transfers)
-		}
-
-		query := `select transaction_id from transfers where transfer_id = ?`
-		stmt, err := db.Prepare(query)
-		if err != nil {
-			t.Fatal(err)
-		}
-		defer stmt.Close()
-
-		var txID string
-		row := stmt.QueryRow(transfers[0].ID)
-		if err := row.Scan(&txID); err != nil {
-			t.Fatal(err)
-		}
-		if txID != transactionID {
-			t.Errorf("incorrect transactionID: %s vs %s", txID, transactionID)
-		}
-	}
-
-	// SQLite tests
-	sqliteDB := database.CreateTestSqliteDB(t)
-	defer sqliteDB.Close()
-	check(t, sqliteDB.DB)
-
-	// MySQL tests
-	mysqlDB := database.CreateTestMySQLDB(t)
-	defer mysqlDB.Close()
-	check(t, mysqlDB.DB)
-}
-
-func TestTransfers__LookupTransferFromReturn(t *testing.T) {
-	t.Parallel()
-
-	check := func(t *testing.T, repo TransferRepository) {
-		amt, _ := model.NewAmount("USD", "32.92")
-		userID := id.User(base.ID())
-		req := &transferRequest{
-			Type:                   model.PushTransfer,
-			Amount:                 *amt,
-			Originator:             model.OriginatorID("originator"),
-			OriginatorDepository:   id.Depository("originator"),
-			Receiver:               model.ReceiverID("receiver"),
-			ReceiverDepository:     id.Depository("receiver"),
-			Description:            "money",
-			StandardEntryClassCode: "PPD",
-			fileID:                 "test-file",
-		}
-		transfers, err := repo.createUserTransfers(userID, []*transferRequest{req})
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		// set metadata after transfer is merged into an ACH file for the FED
-		if err := repo.MarkTransferAsMerged(transfers[0].ID, "merged.ach", "traceNumber"); err != nil {
-			t.Fatal(err)
-		}
-
-		// Now grab the transfer back
-		xfer, err := repo.LookupTransferFromReturn("PPD", amt, "traceNumber", time.Now()) // EffectiveEntryDate is bounded by start and end of a day
-		if err != nil {
-			t.Fatal(err)
-		}
-		if xfer.ID != transfers[0].ID || xfer.UserID != userID.String() {
-			t.Errorf("found other transfer=%q user=(%q vs %q)", xfer.ID, xfer.UserID, userID)
-		}
-	}
-
-	// SQLite tests
-	sqliteDB := database.CreateTestSqliteDB(t)
-	defer sqliteDB.Close()
-	check(t, &SQLTransferRepo{sqliteDB.DB, log.NewNopLogger()})
-
-	// MySQL tests
-	mysqlDB := database.CreateTestMySQLDB(t)
-	defer mysqlDB.Close()
-	check(t, &SQLTransferRepo{mysqlDB.DB, log.NewNopLogger()})
-}
-
-func TestTransfers__SetReturnCode(t *testing.T) {
-	t.Parallel()
-
-	check := func(t *testing.T, db *sql.DB) {
-		userID := id.User(base.ID())
-		returnCode := "R17"
-		amt, _ := model.NewAmount("USD", "51.21")
-
-		repo := &SQLTransferRepo{db, log.NewNopLogger()}
-		requests := []*transferRequest{
-			{
-				Type:                   model.PullTransfer,
-				Amount:                 *amt,
-				Originator:             model.OriginatorID("originator"),
-				OriginatorDepository:   id.Depository("originatorDep"),
-				Receiver:               model.ReceiverID("receiver"),
-				ReceiverDepository:     id.Depository("receiverDep"),
-				Description:            "money2",
-				StandardEntryClassCode: "PPD",
-			},
-		}
-		if _, err := repo.createUserTransfers(userID, requests); err != nil {
-			t.Fatal(err)
-		}
-
-		transfers, err := repo.getUserTransfers(userID)
-		if err != nil || len(transfers) != 1 {
-			t.Errorf("got %d Transfers (error=%v): %v", len(transfers), err, transfers)
-		}
-
-		// Set ReturnCode
-		if err := repo.SetReturnCode(transfers[0].ID, returnCode); err != nil {
-			t.Fatal(err)
-		}
-
-		// Verify
-		transfers, err = repo.getUserTransfers(userID)
-		if err != nil || len(transfers) != 1 {
-			t.Errorf("got %d Transfers (error=%v): %v", len(transfers), err, transfers)
-		}
-		if transfers[0].ReturnCode == nil {
-			t.Fatal("expected ReturnCode")
-		}
-		if transfers[0].ReturnCode.Code != returnCode {
-			t.Errorf("transfers[0].ReturnCode.Code=%s", transfers[0].ReturnCode.Code)
-		}
-
-		t.Logf("%#v", transfers[0].ReturnCode)
-	}
-
-	// SQLite tests
-	sqliteDB := database.CreateTestSqliteDB(t)
-	defer sqliteDB.Close()
-	check(t, sqliteDB.DB)
-
-	// MySQL tests
-	mysqlDB := database.CreateTestMySQLDB(t)
-	defer mysqlDB.Close()
-	check(t, mysqlDB.DB)
 }
 
 func TestTransfers__verifyCustomerStatus(t *testing.T) {
@@ -1532,7 +1107,7 @@ func TestTransfers__getUserEvents(t *testing.T) {
 	depRepo := depository.NewDepositoryRepo(log.NewNopLogger(), db.DB, keeper)
 
 	transferID := id.Transfer(base.ID())
-	transferRepo := &MockTransferRepository{
+	transferRepo := &MockRepository{
 		Xfer: &model.Transfer{
 			ID: transferID,
 		},
